fix mesh upload glbRead from RAM
Former-commit-id: 0829a1a612b089c74db0030f75e08e93d3da099d [formerly 6668c5f3c41d1125712b9d22cf80cd336e658a12]
Former-commit-id: 5974620395274f62ad2cf67e9cb7d016fb0ebe93
Former-commit-id: 56547dc6001b18c5f51e7a37a57da80314f13ce2

+++ src/io/Input.hpp
@@ -9,6 +9,8 @@
 // #include <typeinfo>
 // #include <vector>
 
+#include <cstdlib>
+
 #include "Acquisition.hpp"
 #include "Any.hpp"
 #include "Macros.hpp"
@@ -136,6 +138,10 @@ Container getAll();
     ///
     template <class T>
     void read( std::vector<T>& vector );
+
+//    template <class T>
+//    void readBuffer( s ) {
+//    }
 
     ///
     /// \brief read
@@ -298,7 +304,7 @@ inline void Input::read( std::vector<T>& vector ) {
     std::cout << "[Input] read(std::vector)" << std::endl;
 #endif
 
-    int nbEl;
+    uint64_t nbEl;
     read( nbEl );
 
     vector.clear();
